requirements.txt
Jinja2
beautifulsoup4
python-barcode
pillow
imgkit
qrcode[pil]